Fix duplicates = "closest" for nomatch != NA
If `closest` was called with `length(x) != length(table)` and
`duplicates = "closest", nomatch = 0` (or any nomatch != NA argument) the
warning "longer object length is not a multiple of shorter object length"
was shown.

The primary match result was used to calculate the differences. If
nomatch was non-NA the non-matching elements were accessed, too.

This commit uses NA_integer_ for all non-matching elements and replaces
all NA with `nomatch` in the end. That behaviour is similar to
`base::match`.