fix: move mid to low comparison inside the loop
Before the comparison `mid != low` was in the loop condition. That
results in wrong indices if `x` is larger than `ptable[mid]`. It was
incremented by one (resulting in the correct index), than the condition
was not meet, it was again incremented by one (now incorrectly) before
the condition was tested again.